refactor permission definitions to related resource files
