+++ src/types/accounts.ts
@@ -78,106 +78,6 @@ export const AccountKeyDescriptions: Readonly<{
     "Your account's RSA 2048-bit public key, used to verify signed license files, API responses, and webhooks.",
 } as const
 
-export const UserPermissions = [
-  "account.read",
-  "arch.read",
-  "artifact.read",
-  "channel.read",
-  "component.create",
-  "component.delete",
-  "component.read",
-  "component.update",
-  "constraint.read",
-  "engine.read",
-  "entitlement.read",
-  "group.owners.read",
-  "group.read",
-  "group.users.read",
-  "license.check-in",
-  "license.check-out",
-  "license.create",
-  "license.delete",
-  "license.policy.update",
-  "license.read",
-  "license.renew",
-  "license.revoke",
-  "license.usage.increment",
-  "license.users.attach",
-  "license.users.detach",
-  "license.validate",
-  "machine.check-out",
-  "machine.create",
-  "machine.delete",
-  "machine.heartbeat.ping",
-  "machine.proofs.generate",
-  "machine.read",
-  "machine.update",
-  "package.read",
-  "platform.read",
-  "policy.read",
-  "process.create",
-  "process.delete",
-  "process.heartbeat.ping",
-  "process.read",
-  "process.update",
-  "product.read",
-  "release.download",
-  "release.read",
-  "release.upgrade",
-  "token.generate",
-  "token.read",
-  "token.regenerate",
-  "token.revoke",
-  "user.password.update",
-  "user.read",
-  "user.second-factors.create",
-  "user.second-factors.delete",
-  "user.second-factors.read",
-  "user.second-factors.update",
-  "user.update",
-]
-
-export const LicensePermissions = [
-  "account.read",
-  "arch.read",
-  "artifact.read",
-  "channel.read",
-  "component.read",
-  "constraint.read",
-  "engine.read",
-  "entitlement.read",
-  "group.owners.read",
-  "group.read",
-  "license.check-in",
-  "license.check-out",
-  "license.read",
-  "license.usage.increment",
-  "license.validate",
-  "machine.check-out",
-  "machine.create",
-  "machine.delete",
-  "machine.heartbeat.ping",
-  "machine.proofs.generate",
-  "machine.read",
-  "machine.update",
-  "package.read",
-  "platform.read",
-  "policy.read",
-  "process.create",
-  "process.delete",
-  "process.heartbeat.ping",
-  "process.read",
-  "process.update",
-  "product.read",
-  "release.download",
-  "release.read",
-  "release.upgrade",
-  "token.read",
-  "token.regenerate",
-  "token.revoke",
-  "user.read",
-]
-
 export type AccountSettingAttributes = {
   key: string
   value: string[]
