+++ src/types/licenses.ts
@@ -232,3 +232,44 @@ export type LicenseFilters = {
   machine?: string
   metadata?: Record<string, string>
 }
+
+export const LicensePermissions = [
+  "account.read",
+  "arch.read",
+  "artifact.read",
+  "channel.read",
+  "component.read",
+  "constraint.read",
+  "engine.read",
+  "entitlement.read",
+  "group.owners.read",
+  "group.read",
+  "license.check-in",
+  "license.check-out",
+  "license.read",
+  "license.usage.increment",
+  "license.validate",
+  "machine.check-out",
+  "machine.create",
+  "machine.delete",
+  "machine.heartbeat.ping",
+  "machine.proofs.generate",
+  "machine.read",
+  "machine.update",
+  "package.read",
+  "platform.read",
+  "policy.read",
+  "process.create",
+  "process.delete",
+  "process.heartbeat.ping",
+  "process.read",
+  "process.update",
+  "product.read",
+  "release.download",
+  "release.read",
+  "release.upgrade",
+  "token.read",
+  "token.regenerate",
+  "token.revoke",
+  "user.read",
+]
