+++ src/types/users.ts
@@ -159,3 +159,62 @@ export const InternalRoles: UserRole[] = [
 ]
 
 export const AllRoles: UserRole[] = [...ExternalRoles, ...InternalRoles]
+
+export const UserPermissions = [
+  "account.read",
+  "arch.read",
+  "artifact.read",
+  "channel.read",
+  "component.create",
+  "component.delete",
+  "component.read",
+  "component.update",
+  "constraint.read",
+  "engine.read",
+  "entitlement.read",
+  "group.owners.read",
+  "group.read",
+  "group.users.read",
+  "license.check-in",
+  "license.check-out",
+  "license.create",
+  "license.delete",
+  "license.policy.update",
+  "license.read",
+  "license.renew",
+  "license.revoke",
+  "license.usage.increment",
+  "license.users.attach",
+  "license.users.detach",
+  "license.validate",
+  "machine.check-out",
+  "machine.create",
+  "machine.delete",
+  "machine.heartbeat.ping",
+  "machine.proofs.generate",
+  "machine.read",
+  "machine.update",
+  "package.read",
+  "platform.read",
+  "policy.read",
+  "process.create",
+  "process.delete",
+  "process.heartbeat.ping",
+  "process.read",
+  "process.update",
+  "product.read",
+  "release.download",
+  "release.read",
+  "release.upgrade",
+  "token.generate",
+  "token.read",
+  "token.regenerate",
+  "token.revoke",
+  "user.password.update",
+  "user.read",
+  "user.second-factors.create",
+  "user.second-factors.delete",
+  "user.second-factors.read",
+  "user.second-factors.update",
+  "user.update",
+]
